Move ApplicationPolicyHandler to newdeployer
Not actually needed anymore, but necessary to completely
separate old from new deployer.  The package command uses
this logic in order to generate a policy.  This can be removed
once the SAM template generator is updated to use the new deployer.
Otherwise we need something that can deal with the logic of
the IAM policy loading, which is what the ApplicationPolicyHandler
does.
