+++ chalice/deploy/deployer.py
@@ -3,9 +3,7 @@
 Handles Lambda and API Gateway deployments.
 
 """
-import json
 import sys
-import os
 import textwrap
 import socket
 import uuid
@@ -27,6 +25,7 @@ from chalice.awsclient import TypedAWSClient, ResourceDoesNotExistError
 from chalice.awsclient import DeploymentPackageTooLargeError
 from chalice.awsclient import LambdaClientError
 from chalice.config import Config, DeployedResources  # noqa
+from chalice.deploy.newdeployer import ApplicationPolicyHandler
 from chalice.deploy.packager import LambdaDeploymentPackager
 from chalice.deploy.packager import DependencyBuilder
 from chalice.deploy.swagger import SwaggerGenerator
@@ -861,94 +860,3 @@ class APIGatewayDeployer(object):
                 if function['type'] == 'authorizer':
                     self._aws_client.add_permission_for_authorizer(
                         rest_api_id, function['arn'], str(uuid.uuid4()))
-
-
-class ApplicationPolicyHandler(object):
-    """Manages the IAM policy for an application.
-
-    This class handles returning the policy that used by
-    used for the API handler lambda function for a given
-    stage.
-
-    It has several possible outcomes:
-
-        * By default, it will autogenerate a policy based on
-          analyzing the application source code.
-        * It will return a policy from a file on disk that's been
-          configured as the policy for the given stage.
-
-    This class has a precondition that we should be loading
-    some IAM policy for the the API handler function.
-
-    If a user has indicated that there's a pre-existing
-    role that they'd like to use for the API handler function,
-    this class should never be invoked.  In other words,
-    the logic of whether or not we even need to bother with
-    loading an IAM policy is handled a layer above where
-    this class should be used.
-
-    """
-
-    _EMPTY_POLICY = {
-        'Version': '2012-10-17',
-        'Statement': [],
-    }
-
-    def __init__(self, osutils, policy_generator):
-        # type: (OSUtils, AppPolicyGenerator) -> None
-        self._osutils = osutils
-        self._policy_gen = policy_generator
-
-    def generate_policy_from_app_source(self, config):
-        # type: (Config) -> Dict[str, Any]
-        """Generate a policy from application source code.
-
-        If the ``autogen_policy`` value is set to false, then
-        the .chalice/policy.json file will be used instead of generating
-        the policy from the source code.
-
-        """
-        if config.autogen_policy:
-            app_policy = self._do_generate_from_source(config)
-        else:
-            app_policy = self.load_last_policy(config)
-        return app_policy
-
-    def _do_generate_from_source(self, config):
-        # type: (Config) -> Dict[str, Any]
-        return self._policy_gen.generate_policy(config)
-
-    def load_last_policy(self, config):
-        # type: (Config) -> Dict[str, Any]
-        """Load the last recorded policy file for the app."""
-        filename = self._app_policy_file(config)
-        if config.autogen_policy and not self._osutils.file_exists(filename):
-            return self._EMPTY_POLICY
-        elif not self._osutils.file_exists(filename):
-            raise RuntimeError("Unable to load the policy file. Are you sure "
-                               "it exists?")
-        try:
-            return json.loads(
-                self._osutils.get_file_contents(filename, binary=False)
-            )
-        except ValueError as err:
-            raise RuntimeError("Unable to load the project policy file: %s"
-                               % err)
-
-    def record_policy(self, config, policy_document):
-        # type: (Config, Dict[str, Any]) -> None
-        policy_file = self._app_policy_file(config)
-        policy_json = serialize_to_json(policy_document)
-        self._osutils.set_file_contents(policy_file, policy_json, binary=False)
-
-    def _app_policy_file(self, config):
-        # type: (Config) -> str
-        if config.iam_policy_file:
-            filename = os.path.join(config.project_dir, '.chalice',
-                                    config.iam_policy_file)
-        else:
-            # Otherwise if the user doesn't specify a file it defaults
-            # to a fixed name based on the stage.
-            basename = 'policy-%s.json' % config.chalice_stage
-            filename = os.path.join(config.project_dir, '.chalice', basename)
-        return filename
